+++ package.json
@@ -1,6 +1,5 @@
 {
   "name": "azap",
-  "type": "commonjs",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -29,6 +28,7 @@
     "crypto-js": "^4.2.0",
     "dgram": "^1.0.1",
     "dotenv": "^17.2.3",
+    "express": "^5.2.1",
     "file-saver": "^2.0.5",
     "jose": "^5.2.0",
     "lucide-react": "^0.302.0",
@@ -55,6 +55,7 @@
   },
   "devDependencies": {
     "@types/crypto-js": "^4.2.1",
+    "@types/express": "^5.0.6",
     "@types/file-saver": "^2.0.7",
     "@types/node": "^20",
     "@types/nodemailer": "^6.4.14",
